+++ publish/CHANGELOG.md
@@ -1,4 +1,5 @@
 - v1.32
+  - Adds missing location entries automatically to the `expand_locations.yaml` file.
   - Fixes configs not loading correctly on the first run.
 
 - v1.31